+++ src/index.ts
@@ -2,10 +2,14 @@ import { Readable } from "stream";
 import { FetchError } from "./error";
 import {
 	detectResponseType,
+	FetchEsque,
+	getFetch,
 	isJSONParsable,
 	isJSONSerializable,
 	jsonParse,
 } from "./utils";
+import { FetchSchema, Static } from "./typed";
+import { TNever, TObject } from "@sinclair/typebox";
 
 interface RequestContext {
 	request: Request;
@@ -17,7 +21,7 @@ interface ResponseContext {
 	response: Response;
 }
 
-export interface BetterFetchOptions extends Omit<RequestInit, "body"> {
+export type BetterFetchOptions<B extends Record<string, any> = any> = {
 	/**
 	 * a base url that will be prepended to the url
 	 */
@@ -81,43 +85,49 @@ export interface BetterFetchOptions extends Omit<RequestInit, "body"> {
 	 */
 	duplex?: "full" | "half";
 	/**
-	 * Query parameters
+	 * HTTP method
 	 */
-	query?: Record<string, string | number | boolean | undefined>;
+	method?: PayloadMethod | NonPayloadMethod;
+	/**
+	 * Custom fetch implementation
+	 */
+	customFetchImpl?: FetchEsque;
+	/**
+	 * Plugins
+	 */
+	plugins?: Plugin[];
+} & Omit<RequestInit, "body">;
+
+/**
+ * A plugin that can be used to modify the url and options.
+ * All plugins will be called before the request is made.
+ */
+export interface Plugin {
+	(url: string, options?: FetchOption): Promise<{
+		url: string;
+		options?: FetchOption;
+	}>;
 }
 
 // biome-ignore lint/suspicious/noEmptyInterface: <explanation>
 export interface CreateFetchOption extends BetterFetchOptions {}
 
-export type FetchOption<T extends Record<string, unknown> = any> = (
-	| {
-			body?: never;
-	  }
-	| {
-			method: "POST";
-			body: T;
-	  }
-	| {
-			method: "GET";
-			body?: never;
-	  }
-	| {
-			method:
-				| "PUT"
-				| "DELETE"
-				| "PATCH"
-				| "HEAD"
-				| "OPTIONS"
-				| "CONNECT"
-				| "TRACE";
-			body?: T;
-	  }
-) &
-	BetterFetchOptions;
+export type PayloadMethod = "GET" | "POST" | "PUT" | "PATCH" | "DELETE";
+export type NonPayloadMethod = "GET" | "HEAD" | "OPTIONS";
+
+export type FetchOption<
+	T extends Record<string, unknown> = any,
+	Q extends Record<string, unknown> = any,
+> = InferBody<T> & InferQuery<Q> & BetterFetchOptions;
+
+type InferBody<T> = T extends Record<string, any> ? { body: T } : { body?: T };
+type InferQuery<Q> = Q extends Record<string, any>
+	? { query: Q }
+	: { query?: Q };
 
 export type BetterFetchResponse<
 	T,
-	E extends Record<string, unknown> | unknown,
+	E extends Record<string, unknown> | unknown = unknown,
 > =
 	| {
 			data: T;
@@ -133,19 +143,27 @@ export type BetterFetchResponse<
 	  };
 
 export const betterFetch: BetterFetch = async (url, options) => {
+	const fetch = getFetch(options?.customFetchImpl);
 	const controller = new AbortController();
 	const signal = controller.signal;
 
-	const _url = new URL(`${options?.baseURL ?? ""}${url}`);
+	//run plugins first
+	// const fetcher = createFetch(options);
+	for (const plugin of options?.plugins || []) {
+		const pluginRes = await plugin(url.toString(), options);
+		url = pluginRes.url;
+		options = pluginRes.options;
+	}
+
+	const _url = new URL(`${options?.baseURL ?? ""}${url.toString()}`);
 	const headers = new Headers(options?.headers);
 
 	const shouldStringifyBody =
 		options?.body &&
-		!(options.body instanceof FormData) &&
-		!(options.body instanceof URLSearchParams) &&
 		isJSONSerializable(options.body) &&
 		(!headers.has("content-type") ||
-			headers.get("content-type") === "application/json");
+			headers.get("content-type") === "application/json") &&
+		typeof options?.body !== "string";
 
 	if (shouldStringifyBody) {
 		!headers.has("content-type") &&
@@ -162,11 +180,16 @@ export const betterFetch: BetterFetch = async (url, options) => {
 		signal,
 		...options,
 		body: shouldStringifyBody
-			? JSON.stringify(options.body)
+			? JSON.stringify(options?.body)
 			: options?.body
 			? options.body
 			: undefined,
 		headers,
+		method: options?.method?.length
+			? options.method
+			: options?.body
+			? "POST"
+			: "GET",
 	};
 
 	if (
@@ -195,7 +218,9 @@ export const betterFetch: BetterFetch = async (url, options) => {
 	}
 
 	await options?.onRequest?.(context);
-	const response = await fetch(context.request);
+
+	const response = await fetch(_url.toString(), _options);
+
 	const responseContext: ResponseContext = {
 		response,
 	};
@@ -259,6 +284,7 @@ export const betterFetch: BetterFetch = async (url, options) => {
 			},
 		};
 	}
+
 	return {
 		data: null,
 		error: {
@@ -269,9 +295,13 @@ export const betterFetch: BetterFetch = async (url, options) => {
 	};
 };
 
-export const createFetch = <R = unknown, F = unknown>(
+export const createFetch = <
+	Routes extends FetchSchema = any,
+	R = unknown,
+	E = unknown,
+>(
 	config?: CreateFetchOption,
-): BetterFetch<R, F> => {
+): BetterFetch<Routes, R, E> => {
 	const $fetch: BetterFetch = async (url, options) => {
 		return await betterFetch(url, {
 			...config,
@@ -284,11 +314,37 @@ export const createFetch = <R = unknown, F = unknown>(
 
 betterFetch.native = fetch;
 
-export interface BetterFetch<BaseT = any, BaseE = unknown> {
-	<T = BaseT, E = BaseE>(url: string | URL, options?: FetchOption): Promise<
-		BetterFetchResponse<T, E>
+export interface BetterFetch<
+	Routes extends FetchSchema = {
+		[key in string]: {
+			output: TNever;
+		};
+	},
+	BaseT = any,
+	BaseE = unknown,
+> {
+	<T = BaseT, E = BaseE, K extends keyof Routes = keyof Routes>(
+		url: K | URL | Omit<string, keyof Routes>,
+		...options: Routes[K]["input"] extends TObject
+			? [
+					FetchOption<
+						Static<Routes[K]["input"]>,
+						Routes[K]["query"] extends TObject
+							? Static<Routes[K]["query"]>
+							: any
+					>,
+			  ]
+			: Routes[K]["query"] extends TObject
+			? [FetchOption<any, Static<Routes[K]["query"]>>]
+			: [FetchOption?]
+	): Promise<
+		BetterFetchResponse<
+			Routes[K]["output"] extends TObject ? Static<Routes[K]["output"]> : T,
+			E
+		>
 	>;
 	native: typeof fetch;
 }
+
 export type CreateFetch = typeof createFetch;
 export default betterFetch;